access-api/requirements.txt
aioredis==2.0.1
fastapi==0.95.2
sqlalchemy==2.0.13
sqlalchemy_utils==0.41.1
orjson==3.8.12
pydantic==1.10.7
uvicorn==0.20.0
uvloop==0.17.0
python-dotenv==0.21.1
sentry-sdk[fastapi]==1.23.0
gunicorn==20.1.0
